MAINT: allocate namespace and pass to test runner pod
This patch focusses on passing down the generated namespace to the test
runner pod. This allows for the test runner to run in the same namespace
as the test scenario itself, which means we will not need any RBAC config
to allow the test runner to access the test scenario resources.